% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{apipop}
\alias{apipop}
\title{apipop Dataset}
\format{
Data frame: 6,194 rows, 37 variables.
}
\source{
Academic Performance Index, provided via the survey package (Thomas Lumley).
}
\usage{
apipop
}
\description{
Student performance in california schools.
}
\keyword{datasets}
