% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CASchools}
\alias{CASchools}
\title{CASchools Dataset}
\format{
Data frame with district-level educational data.
}
\source{
Extracted from the AER package (Christian Kleiber & Achim Zeileis), based on Stock and Watson.
}
\usage{
CASchools
}
\description{
California test score data.
}
\keyword{datasets}
