% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CollegeDistance}
\alias{CollegeDistance}
\title{CollegeDistance Dataset}
\format{
Data frame: 4,739 rows, 14 variables.
}
\usage{
CollegeDistance
}
\description{
College distance data from the High School and Beyond survey.
}
\keyword{datasets}
