% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EducationLiteracy}
\alias{EducationLiteracy}
\title{EducationLiteracy Dataset}
\format{
Data frame: 193 rows, 4 variables.
}
\usage{
EducationLiteracy
}
\description{
Education spending and literacy rates for countries.
}
\keyword{datasets}
