% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ExamScores}
\alias{ExamScores}
\title{ExamScores Dataset}
\format{
Data frame: 4,059 rows, 10 variables.
}
\usage{
ExamScores
}
\description{
Exam results for six inner London Education Authorities.
}
\keyword{datasets}
