% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FirstYearGPA}
\alias{FirstYearGPA}
\title{FirstYearGPA Dataset}
\format{
Data frame: 219 rows, 10 variables.
}
\usage{
FirstYearGPA
}
\description{
Predicting first-year college GPA for students.
}
\keyword{datasets}
