% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pisausa}
\alias{pisausa}
\title{pisausa Dataset}
\format{
Data frame with assessment scores and demographics.
}
\source{
OECD Programme for International Student Assessment (PISA).
}
\usage{
pisausa
}
\description{
International student assessment data for the USA.
}
\keyword{datasets}
