% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{school}
\alias{school}
\title{school Dataset}
\format{
Data frame with institutional data.
}
\source{
U.S. Department of Education (College Scorecard metrics).
}
\usage{
school
}
\description{
Institutional metrics and school performance indicators.
}
\keyword{datasets}
