% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{schoolProgram}
\alias{schoolProgram}
\title{schoolProgram Dataset}
\format{
Data frame with program characteristics.
}
\source{
U.S. Department of Education.
}
\usage{
schoolProgram
}
\description{
Educational programs and institutional offerings data.
}
\keyword{datasets}
