% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{STARplus}
\alias{STARplus}
\title{STARplus Dataset}
\format{
Data frame: 13,381 rows, 56 variables.
}
\source{
Project STAR, primary investigator Jeremy Finn.
}
\usage{
STARplus
}
\description{
STAR participants plus nonexperimental controls.
}
\keyword{datasets}
