% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{STUDENT}
\alias{STUDENT}
\title{STUDENT Dataset}
\format{
Data frame: 607 rows, 19 variables.
}
\usage{
STUDENT
}
\description{
College GPAs of students in an introductory statistics class.
}
\keyword{datasets}
