% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CASchools}
\alias{CASchools}
\title{CASchools Dataset}
\format{
Data frame: 420 rows, 14 variables.
}
\usage{
CASchools
}
\description{
California Test Score Data on school performance and demographics.
}
\keyword{datasets}
