% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{classroom}
\alias{classroom}
\title{classroom Dataset}
\format{
Data frame: 1,190 rows, 12 variables.
}
\source{
Study of Instructional Improvement, research contributor Andrew Raudenbush.
}
\usage{
classroom
}
\description{
classroom data in Chapter 4 of the study of Instructional Improvement.
}
\keyword{datasets}
