% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EducationLiteracy}
\alias{EducationLiteracy}
\title{EducationLiteracy Dataset}
\format{
Data frame with social science indicators.
}
\source{
Extracted from the carData package (John Fox).
}
\usage{
EducationLiteracy
}
\description{
Education and literacy survey metrics.
}
\keyword{datasets}
