% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ExamScores}
\alias{ExamScores}
\title{ExamScores Dataset}
\format{
Data frame with academic scores and variables.
}
\source{
Extracted from the carData package (John Fox).
}
\usage{
ExamScores
}
\description{
Student exam performance metrics.
}
\keyword{datasets}
