% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FirstYearGPA}
\alias{FirstYearGPA}
\title{FirstYearGPA Dataset}
\format{
Data frame with student performance metrics.
}
\source{
Extracted from the carData package (John Fox).
}
\usage{
FirstYearGPA
}
\description{
Academic performance data tracking first year university GPA.
}
\keyword{datasets}
