% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MathAchieve}
\alias{MathAchieve}
\title{MathAchieve Dataset}
\format{
Data frame: 7,185 rows, 6 variables.
}
\usage{
MathAchieve
}
\description{
Mathematics achievement scores.
}
\keyword{datasets}
