% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NELS}
\alias{NELS}
\title{NELS Dataset}
\format{
Data frame: 500 rows, 48 variables.
}
\source{
National Center for Education Statistics (NCES).
}
\usage{
NELS
}
\description{
National Education Longitudinal study (NELS) of 1988.
}
\keyword{datasets}
