% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pisausa}
\alias{pisausa}
\title{pisausa Dataset}
\format{
Data frame: 5,233 rows, 65 variables.
}
\usage{
pisausa
}
\description{
2009 PISA results for students in the United States.
}
\keyword{datasets}
