% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{school}
\alias{school}
\title{school Dataset (CollegeScorecard)}
\format{
Data frame: 11,300 rows, 25 variables.
}
\usage{
school
}
\description{
Information about institutions in the U.S. College Scorecard.
}
\keyword{datasets}
