% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{schoolProgram}
\alias{schoolProgram}
\title{schoolProgram Dataset}
\format{
Data frame: 200 rows, 11 variables.
}
\usage{
schoolProgram
}
\description{
High School and Beyond Survey data on student programs.
}
\keyword{datasets}
