% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{schools}
\alias{schools}
\title{schools Dataset}
\format{
Data frame: 260 rows, 19 variables.
}
\usage{
schools
}
\description{
Schools data set (NELS-88) for multilevel modeling.
}
\keyword{datasets}
