% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{STUDENT}
\alias{STUDENT}
\title{STUDENT Dataset}
\format{
Data frame with applied statistics subsets.
}
\source{
OpenIntro Statistics educational data repository.
}
\usage{
STUDENT
}
\description{
Student survey metrics and statistics.
}
\keyword{datasets}
